requirements.txt
Versions:

selenium~=4.25.0
pytest~=8.3.3
Python~=3.12.6
Chrome~=131.0.6778.86
PyYAML~=6.0.2